Message

+++ exec7.c
@@ -3,8 +3,10 @@
 int main() {
     int distance_km = 100;
     int vitesse_kmh = 50;
-    int temps = distance_km / vitesse_kmh; 
+    float temps = distance_km / vitesse_kmh; 
     int minutes = temps * 60;
     printf("minutes = %d\n", minutes); 
     return 0;
 }
+
+/*il est mieux de définir le type de valeur d'une division comme float*/
